// Nursing.Svelte/src/lib/api/models/FeedingDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Nursing API
 * API for Nursing App
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface FeedingDto
 */
export interface FeedingDto {
    /**
     * 
     * @type {string}
     * @memberof FeedingDto
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof FeedingDto
     */
    leftBreastTotal?: string;
    /**
     * 
     * @type {string}
     * @memberof FeedingDto
     */
    rightBreastTotal?: string;
    /**
     * 
     * @type {string}
     * @memberof FeedingDto
     */
    totalTime?: string;
    /**
     * 
     * @type {Date}
     * @memberof FeedingDto
     */
    started?: Date;
    /**
     * 
     * @type {Date}
     * @memberof FeedingDto
     */
    finished?: Date | null;
    /**
     * 
     * @type {boolean}
     * @memberof FeedingDto
     */
    lastIsLeft?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof FeedingDto
     */
    lastUpdated?: Date;
    /**
     * 
     * @type {Date}
     * @memberof FeedingDto
     */
    deleted?: Date | null;
}

/**
 * Check if a given object implements the FeedingDto interface.
 */
export function instanceOfFeedingDto(value: object): value is FeedingDto {
    return true;
}

export function FeedingDtoFromJSON(json: any): FeedingDto {
    return FeedingDtoFromJSONTyped(json, false);
}

export function FeedingDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): FeedingDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'leftBreastTotal': json['leftBreastTotal'] == null ? undefined : json['leftBreastTotal'],
        'rightBreastTotal': json['rightBreastTotal'] == null ? undefined : json['rightBreastTotal'],
        'totalTime': json['totalTime'] == null ? undefined : json['totalTime'],
        'started': json['started'] == null ? undefined : (new Date(json['started'])),
        'finished': json['finished'] == null ? undefined : (new Date(json['finished'])),
        'lastIsLeft': json['lastIsLeft'] == null ? undefined : json['lastIsLeft'],
        'lastUpdated': json['lastUpdated'] == null ? undefined : (new Date(json['lastUpdated'])),
        'deleted': json['deleted'] == null ? undefined : (new Date(json['deleted'])),
    };
}

export function FeedingDtoToJSON(json: any): FeedingDto {
    return FeedingDtoToJSONTyped(json, false);
}

export function FeedingDtoToJSONTyped(value?: FeedingDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'leftBreastTotal': value['leftBreastTotal'],
        'rightBreastTotal': value['rightBreastTotal'],
        'totalTime': value['totalTime'],
        'started': value['started'] == null ? undefined : ((value['started']).toISOString()),
        'finished': value['finished'] == null ? undefined : ((value['finished'] as any).toISOString()),
        'lastIsLeft': value['lastIsLeft'],
        'lastUpdated': value['lastUpdated'] == null ? undefined : ((value['lastUpdated']).toISOString()),
        'deleted': value['deleted'] == null ? undefined : ((value['deleted'] as any).toISOString()),
    };
}