// Nursing.Svelte/src/lib/api/models/SyncResult.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Nursing API
 * API for Nursing App
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { FeedingDto } from './FeedingDto';
import {
    FeedingDtoFromJSON,
    FeedingDtoFromJSONTyped,
    FeedingDtoToJSON,
    FeedingDtoToJSONTyped,
} from './FeedingDto';

/**
 * 
 * @export
 * @interface SyncResult
 */
export interface SyncResult {
    /**
     * 
     * @type {boolean}
     * @memberof SyncResult
     */
    success?: boolean;
    /**
     * 
     * @type {Array<FeedingDto>}
     * @memberof SyncResult
     */
    feedings: Array<FeedingDto>;
    /**
     * 
     * @type {Array<string>}
     * @memberof SyncResult
     */
    badIds: Array<string>;
    /**
     * 
     * @type {number}
     * @memberof SyncResult
     */
    updates?: number;
}

/**
 * Check if a given object implements the SyncResult interface.
 */
export function instanceOfSyncResult(value: object): value is SyncResult {
    if (!('feedings' in value) || value['feedings'] === undefined) return false;
    if (!('badIds' in value) || value['badIds'] === undefined) return false;
    return true;
}

export function SyncResultFromJSON(json: any): SyncResult {
    return SyncResultFromJSONTyped(json, false);
}

export function SyncResultFromJSONTyped(json: any, ignoreDiscriminator: boolean): SyncResult {
    if (json == null) {
        return json;
    }
    return {
        
        'success': json['success'] == null ? undefined : json['success'],
        'feedings': ((json['feedings'] as Array<any>).map(FeedingDtoFromJSON)),
        'badIds': json['badIds'],
        'updates': json['updates'] == null ? undefined : json['updates'],
    };
}

export function SyncResultToJSON(json: any): SyncResult {
    return SyncResultToJSONTyped(json, false);
}

export function SyncResultToJSONTyped(value?: SyncResult | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'success': value['success'],
        'feedings': ((value['feedings'] as Array<any>).map(FeedingDtoToJSON)),
        'badIds': value['badIds'],
        'updates': value['updates'],
    };
}